refactor(catplus-common): make models structure flat
