+++ src/catplus-common/src/models/hci.rs
