+++ src/catplus-common/src/models/mod.rs
@@ -1,10 +1,12 @@
 pub mod agilent;
 pub mod core;
 pub mod enums;
-pub mod types;
+pub mod hci;
+pub mod synth;
 
 // Re-export all models;
 pub use agilent::*;
 pub use core::*;
 pub use enums::*;
-pub use types::*;
+pub use hci::*;
+pub use synth::*;
