--- src/catplus-common/src/models/types/mod.rs
@@ -1,2 +1,0 @@
-pub mod hci;
-pub mod synth;
